pyproject.toml
[project]
name = "dev-to-digest"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "rmc-8.com", email = "k@rmc-8.com" }
]
dependencies = [
    "devtopy>=0.1.1",
    "langchain>=0.2.12",
    "requests>=2.32.3",
    "fire>=0.6.0",
    "python-dotenv>=1.0.1",
    "langchain-community>=0.2.11",
    "langchain-openai>=0.1.20",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dev_to_digest"]
